+++ wails/frontend/wailsjs/go/models.ts
@@ -1,0 +1,19 @@
+export namespace main {
+	
+	export class Printer {
+	    name: string;
+	    displayName: string;
+	
+	    static createFrom(source: any = {}) {
+	        return new Printer(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.name = source["name"];
+	        this.displayName = source["displayName"];
+	    }
+	}
+
+}
+